+++ src/createRoom.ts
@@ -42,7 +42,8 @@ import ErrorDialog from "./components/views/dialogs/ErrorDialog";
 import Spinner from "./components/views/elements/Spinner";
 import { ViewRoomPayload } from "./dispatcher/payloads/ViewRoomPayload";
 import { findDMForUser } from "./utils/dm/findDMForUser";
-import { privateShouldBeEncrypted, shouldForceDisableEncryption } from "./utils/rooms";
+import { privateShouldBeEncrypted } from "./utils/rooms";
+import { shouldForceDisableEncryption } from "./utils/room/shouldForceDisableEncryption";
 import { waitForMember } from "./utils/membership";
 import { PreferredRoomVersions } from "./utils/PreferredRoomVersions";
 import SettingsStore from "./settings/SettingsStore";
